requirements.txt
Flask==2.3.3
numpy==1.24.3
opencv-python==4.8.0.74
keras==2.12.0
Pillow==9.5.0

#  pip install -r requirements.txt